backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
openai==1.10.0
pydantic==2.6.0
python-dotenv==1.0.1
httpx==0.27.2
motor>=3.6.0
beanie>=1.25.0
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-jose[cryptography]
